+++ fw/src/net/sourceforge/squirrel_sql/fw/datasetviewer/cellcomponent/DataTypeShort.java
@@ -115,8 +115,8 @@ public class DataTypeShort
 	 * Determine if two objects of this data type contain the same value.
 	 * Neither of the objects is null
 	 */
-	public boolean areEqual(Object obj1, Object obj2) {
-		return ((Short)obj1).equals(obj2);
+	public boolean areEqual(Object obj1, Object obj2){
+		return (obj1).equals(obj2);
 	}
 
 	/*
